hr-assistant/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
openai
pymongo
motor
streamlit
python-dotenv
aiofiles